requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
pydantic
pydantic-settings
openai
crewai
python-dotenv
langchain
langchain-community
